Relocate user profile to distinct namespace
