+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/CCDLibAutoConfigure.java
@@ -17,12 +17,12 @@ import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import uk.gov.hmcts.ccd.AliasWebConfig;
 import uk.gov.hmcts.ccd.CoreCaseDataApplication;
-import uk.gov.hmcts.ccd.UserProfileApplication;
+import uk.gov.hmcts.ccd.userprofile.UserProfileApplication;
 import uk.gov.hmcts.ccd.data.AuthClientsConfiguration;
 import uk.gov.hmcts.ccd.definition.store.CaseDataAPIApplication;
 import uk.gov.hmcts.ccd.definition.store.repository.AuthClientConfiguration;
 import uk.gov.hmcts.ccd.security.JwtGrantedAuthoritiesConverter;
-import uk.gov.hmcts.ccd.hikari.HikariConfigurationPropertiesReportEndpoint;
+import uk.gov.hmcts.ccd.userprofile.hikari.HikariConfigurationPropertiesReportEndpoint;
 import uk.gov.hmcts.reform.authorisation.validators.ServiceAuthTokenValidator;
 import uk.gov.hmcts.reform.idam.client.IdamApi;
 import uk.gov.hmcts.reform.roleassignment.RoleAssignmentApplication;
