+++ settings.gradle
@@ -11,3 +11,4 @@ includeBuild 'ccd-user-profile-api'
 includeBuild 'am-role-assignment-service'
 include 'rse-cft-lib'
 include 'lib-consumer'
+include 'ccd-user-profile-api-relocated'
